simple checker for OS

+++ tensorflow/issue_62549/reproduce_bug.sh
@@ -1,2 +1,7 @@
+if [[ "$OSTYPE" != "linux-gnu"* ]]; then
+    echo "Broken assumption: Script was tested only on Linux."
+    exit
+fi
+
 docker build -t issue_62549 .
 docker run -it --rm --gpus all issue_62549
